+++ packages/core/src/index.ts
@@ -30,3 +30,4 @@ export * from "./compare.js";
 export * from "./check/index.js";
 export * from "./transcript/index.js";
 export * from "./link.js";
+export * from "./report.js";